Fix Amiibos in Amiibo Info Stats and support image
Relocate inline BaseAdapter to individual class file

Respect cache settings for Amiibo Info Stats image
